requirements.txt
fastapi 
transformers
uvicorn 
torch 
torchvision 
pillow 
requests
cryptography
